fix: linting error

+++ medmodels/treatment_effect/matching/tests/test_metrics.py
@@ -25,13 +25,13 @@ class TestMetrics(unittest.TestCase):
                 [29, 33, 37, 46, 55],
             ]
         )
-        inv_cov = np.linalg.inv(np.cov(data))
+        inv_cov = np.linalg.inv(np.cov(data)).astype(np.float64)
         a1, a2 = np.array([68, 600, 40]), np.array([66, 640, 44])
         result = metrics.mahalanobis_metric(a1, a2, inv_cov=inv_cov)
         assert result == pytest.approx(5.33, 2)
 
         data = np.array([[-2.1, -1, 4.3]])
-        inv_cov = 1 / np.cov(data)
+        inv_cov = (1 / np.cov(data)).astype(np.float64)
         a1, a2 = np.array([1]), np.array([2])
         result = metrics.mahalanobis_metric(a1, a2, inv_cov=inv_cov)
         assert result == pytest.approx(0.29, 2)
